Introduce BaseTestRepository
