+++ app/src/test/java/com/book/library/repository/SchemaValidationTest.java
@@ -1,13 +1,8 @@
 package com.book.library.repository;
 
-import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
-@DataJpaTest(properties = "spring.jpa.hibernate.ddl-auto=validate")
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-class SchemaValidationTest implements WithAssertions {
+class SchemaValidationTest extends BaseTestRepository {
 
     @Test
     void testSchemaValidity() {}
